feat: 특정 시간에 대한 예약이 존재할 때 예약 시간을 삭제할 수 없다

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import org.springframework.stereotype.Service;
 import roomescape.controller.dto.ReservationTimeRequest;
 import roomescape.controller.dto.ReservationTimeResponse;
+import roomescape.repository.ReservationDao;
 import roomescape.repository.ReservationTimeDao;
 import roomescape.service.reservation.ReservationTime;
 
@@ -11,13 +12,15 @@ import roomescape.service.reservation.ReservationTime;
 public class ReservationTimeService {
 
     private final ReservationTimeDao reservationTimeDao;
+    private final ReservationDao reservationDao;
 
-    public ReservationTimeService(final ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(final ReservationTimeDao reservationTimeDao, final ReservationDao reservationDao) {
         this.reservationTimeDao = reservationTimeDao;
+        this.reservationDao = reservationDao;
     }
 
     public ReservationTimeResponse createReservationTime(final ReservationTimeRequest reservationTimeRequest) {
-        if (reservationTimeDao.isExists(reservationTimeRequest.startAt())) {
+        if (reservationTimeDao.isExistsByTime(reservationTimeRequest.startAt())) {
             throw new IllegalArgumentException("이미 존재하는 시간입니다.");
         }
         final ReservationTime convertedRequest = reservationTimeRequest.converToReservationTime();
@@ -32,6 +35,9 @@ public class ReservationTimeService {
     }
 
     public void deleteReservationTime(final long id) {
+        if (reservationDao.isExistsByTimeId(id)) {
+            throw new IllegalArgumentException("예약이 존재하는 시간은 삭제할 수 없습니다.");
+        }
         reservationTimeDao.deleteReservationTimeById(id);
     }
 }
